requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.57.0
frozenlist==1.5.0
geoip2==5.0.1
idna==3.10
kiwisolver==1.4.8
MarkupSafe==3.0.2

# Added to support Streamlit dashboard + reports + CVE mapping
streamlit==1.32.2
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.3
openpyxl==3.1.2
requests==2.31.0

